refactor(memory): route archive into memory captures
Memory is the product surface for capture review, so legacy
archive links now redirect into the Memory captures queue with
query parameters preserved.

Moved RawCaptureReview behavior coverage to the component and kept
ArchivePage tests focused on the redirect contract.
